+++ MetriCam2/TypeConversion.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Globalization;
+using Metrilus.Util;
+
+namespace MetriCam2
+{
+    class TypeConversion
+    {
+        public static string GetAsGoodString(object value)
+        {
+            return GetAsGoodString(value, value.GetType());
+        }
+
+        public static string GetAsGoodString(object value, Type valueType)
+        {
+            string valueAsString;
+            if (valueType == typeof(float))
+            {
+                valueAsString = ((float)value).ToString("R", CultureInfo.InvariantCulture);
+            }
+            else if (valueType == typeof(double))
+            {
+                valueAsString = ((double)value).ToString("R", CultureInfo.InvariantCulture);
+            }
+            else if (valueType == typeof(Point2i))
+            {
+                valueAsString = string.Format("{0}x{1}", ((Point2i)value).X, ((Point2i)value).Y);
+            }
+            else
+            {
+                valueAsString = value.ToString();
+            }
+            return valueAsString;
+        }
+    }
+}
